binance_spot_signals_ws/requirements.txt

httpx==0.27.2
websockets==12.0
PyYAML==6.0.2
numpy==1.26.4
pydantic==2.8.2
